https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved sendEmail services to esn.inbox.libs
