+++ src/esn.inbox.libs/app/app.module.js
@@ -18,3 +18,5 @@ require('./services/email-body/email-body');
 require('./services/jmap-client-wrapper/jmap-client-wrapper.service');
 require('./services/action/background-action.service');
 require('./services/action/async-jmap-action.service');
+require('./services/send-email/email-sending.service');
+require('./services/send-email/send-email.service');
